Renaming 'spring' mazerunner scenarios to 'spring boot'
