+++ features/fixtures/mazerunnerspringboot/src/main/java/com/bugsnag/mazerunnerspringboot/Application.java
@@ -1,4 +1,4 @@
-package com.bugsnag.mazerunnerspring;
+package com.bugsnag.mazerunnerspringboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
