+++ features/fixtures/mazerunnerspringboot/src/main/java/com/bugsnag/mazerunnerspringboot/TestRestController.java
@@ -1,4 +1,4 @@
-package com.bugsnag.mazerunnerspring;
+package com.bugsnag.mazerunnerspringboot;
 
 import com.bugsnag.Bugsnag;
 import org.apache.log4j.Logger;
