+++ features/steps/build_steps.rb
@@ -17,11 +17,11 @@ When("I run {string} with logback config {string}") do |eventType, logback_confi
   }
 end
 
-When("I run spring {string} with the defaults") do |eventType|
+When("I run spring boot {string} with the defaults") do |eventType|
   steps %Q{
     And I set environment variable "MOCK_API_PATH" to "http://localhost:$MOCK_API_PORT"
     And I set environment variable "BUGSNAG_API_KEY" to "a35a2a72bd230ac0aa0f52715bbdc6aa"
     And I set environment variable "EVENT_TYPE" to "#{eventType}"
-    And I run the script "features/scripts/run-java-spring-app.sh" synchronously
+    And I run the script "features/scripts/run-java-spring-boot-app.sh" synchronously
   }
 end
